04/15 database

+++ src/main/java/com/sidam_backend/data/WorkAlarm.java
@@ -16,7 +16,7 @@ public class WorkAlarm implements Serializable {
     private Long id;
 
     @NotNull
-    private long time;
+    private int time;
     // 분 단위
 
     @ManyToOne
